+++ python/sglang/srt/model_executor/cuda_graph_runner.py
@@ -34,7 +34,7 @@ from sglang.srt.managers.schedule_batch import (
     ScheduleBatch,
     init_flashinfer_args,
 )
-from sglang.srt.model_executor.forward_batch import ForwardMode
+from sglang.srt.model_executor.forward_batch_info import ForwardMode
 from sglang.srt.utils import monkey_patch_vllm_all_gather
