+++ python/sglang/srt/model_executor/forward_batch_info.py
@@ -12,7 +12,7 @@ class ForwardMode(IntEnum):
 
 
 @dataclass
-class ForwardBatch:
+class ForwardBatchInfo:
     """Model forward information for a batch."""
 
     pass
